refactor: Consolidate editor toolbars into a global EditorToolbar for improved consistency and functionality across all editors
